[RELEASE.0.1]_INPC Bug w/ Spawner, Cam & Extrovert Bug, & Dust Particle System
INPC's wait state bug fixed,
Surprise INPC's spawn area is limited to 90 degrees in front of the PC.
The bug of starting the level while the camera is in motion has been resolved.
When you start the game one after the other from the editor, the score is not reset, it has been solved.
Running effect with particles has been made for pc's.
RELEASE.0.1