Coin Data, Power Up, Gun Spawn, & Collectable Models
Collectable models and a gun model have imported to the project.
Coin, Bullet, Gun, and Surprise Box spawner scripts have created.
An UI example has created for power up information.
Coin count is stored in CollectableData for use of UI etc.
Issues:
IsLearning bool is controlled in StartPanel which is a UI script, because of the coroutine usage.
Going to work on:
Power Up UI's animation,
Shock state.